just a little touchs

+++ app/Http/Controllers/Dashboard/CategoryController.php
@@ -8,6 +8,8 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Category\StoreRequest;
 use App\Http\Requests\Category\UpdateRequest;
 use Illuminate\Support\Facades\Auth;
+use App\Exports\CategoriesExport;
+use Maatwebsite\Excel\Facades\Excel;
 
 class CategoryController extends Controller
 {
@@ -15,9 +17,12 @@ class CategoryController extends Controller
     public function index()
     {
         // view all the category
-        if((Auth::user())->hasRole('admin')){
-        $categories = Category::select('*')->paginate(10);
-        return view('dashboard.categories.index')->with('categories',$categories);
+        if((Auth::user())->ability('admin|dataEntry|readOnly', 'categories-read')){
+            $categories = Category::select('*')->paginate(10);
+            return view('dashboard.categories.index')->with('categories',$categories);
+        }
+        else {
+           return view('dashboard.permission_denied');
         }
     }
 
@@ -25,7 +30,13 @@ class CategoryController extends Controller
     public function create()
     {
         // create new category
-        return view('dashboard.categories.create');
+        if((Auth::user())->ability('admin|dataEntry', 'categories-create')){
+
+            return view('dashboard.categories.create');
+        }
+        else {
+            return view('dashboard.permission_denied');
+        }
     }
 
 
@@ -46,12 +57,18 @@ class CategoryController extends Controller
 
     public function edit($id)
     {
-         // edit category
-         $category = Category::find($id);
-         if($category == null){
-            return redirect()->back();
+        // edit category
+        if((Auth::user())->ability('admin|dataEntry', 'categories-edit')){
+
+            $category = Category::find($id);
+            if($category == null){
+                return redirect()->back();
+            }
+            return view('dashboard.categories.edit')->with('category',$category);
         }
-        return view('dashboard.categories.edit')->with('category',$category);
+        else {
+            return view('dashboard.permission_denied');
+        }
     }
 
 
@@ -76,11 +93,22 @@ class CategoryController extends Controller
     public function destroy($id)
     {
         // delete category
-        $category = Category::find($id);
-        if($category == null){
-            return redirect()->back()->withErrors("There is no such category");
+        if((Auth::user())->ability('admin|dataEntry', 'categories-delete')){
+
+            $category = Category::find($id);
+            if($category == null){
+                return redirect()->back()->withErrors("There is no such category");
+            }
+            $category->delete($id);
+            return redirect()->back();
         }
-        $category->delete($id);
-        return redirect()->back();
+        else {
+            return view('dashboard.permission_denied');
+        }
+    }
+
+    public function export()
+    {
+        return Excel::download(new CategoriesExport , 'Categories.xlsx');
     }
 }
